requirements.txt
fastapi
deta
jinja2
python-dotenv
requests